+++ src/gui/preferences_window.rs
@@ -1,6 +1,7 @@
 use {
     super::{window_open::WindowOpen, Gui},
     crate::{app::App, config},
+    egui_sfml::sfml::graphics::RenderWindow,
 };
 
 #[derive(Default)]
@@ -9,7 +10,8 @@ pub struct PreferencesWindow {
 }
 
 impl PreferencesWindow {
-    pub fn ui(ui: &mut egui::Ui, _gui: &mut Gui, app: &mut App) {
+    pub fn ui(ui: &mut egui::Ui, _gui: &mut Gui, app: &mut App, rwin: &mut RenderWindow) {
+        ui.heading("Style");
         let style = &mut app.cfg.style;
         ui.heading("Font sizes");
         let mut any_changed = false;
@@ -50,5 +52,17 @@ impl PreferencesWindow {
         if any_changed {
             crate::gui::set_font_sizes_ctx(ui.ctx(), style);
         }
+        ui.separator();
+        ui.heading("Video");
+        if ui.checkbox(&mut app.cfg.vsync, "Vsync").clicked() {
+            rwin.set_vertical_sync_enabled(app.cfg.vsync);
+        }
+        ui.horizontal(|ui| {
+            ui.label("FPS limit (0 to disable)");
+            ui.add(egui::DragValue::new(&mut app.cfg.fps_limit));
+            if ui.button("Set").clicked() {
+                rwin.set_framerate_limit(app.cfg.fps_limit);
+            }
+        });
     }
 }