+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Layout properties is accessible from right click context menu on the layout
 - Error reporting message dialog if the program panics
 - Each file can set a metafile association to always load that meta when loaded
+- Vsync and fps limit settings in preferences window
 
 ### Changed